test1.0

+++ src/main/java/com/Testing/practicasTesteo/service/ArticleServiceImpl.java
@@ -2,11 +2,14 @@ package com.Testing.practicasTesteo.service;
 
 import com.Testing.practicasTesteo.entity.Article;
 import com.Testing.practicasTesteo.exceptions.ArticleNotFoundException;
+import com.Testing.practicasTesteo.exceptions.NotFoundException;
+import com.Testing.practicasTesteo.exceptions.NotSavedException;
 import com.Testing.practicasTesteo.respository.ArticleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -17,34 +20,49 @@ public class ArticleServiceImpl implements ArticleService {
 
 
     @Override
-    public List<Article> getAllArticles() {
-        return articleRepository.findAll();
+    public List<Article> getAllArticles() throws NotFoundException {
+        try{
+            return articleRepository.findAll();
+        }catch (Exception e){
+            throw new ArticleNotFoundException("No data found" + e.getMessage());
+
+        }
+
     }
 
     @Override
-    public Optional<Article> getArticleByid(long id) throws ArticleNotFoundException {
-        return articleRepository.findById(id);
+    public Article getArticleByid(long id) throws ArticleNotFoundException {
+        try {
+            return articleRepository.findById(id).orElseThrow();
+        } catch (NoSuchElementException e) {
+            throw new ArticleNotFoundException("Article not found with id: " + id);
+        }
     }
 
     @Override
-    public Optional<Article> updateArticleById(Article a, long id) {
+    public Article updateArticleById(Article a, long id) throws ArticleNotFoundException {
         Optional<Article> articleFound = articleRepository.findById(id);
-        if (articleFound.isPresent()){
+        if (articleFound.isPresent()) {
             Article articleUpdated = articleFound.get();
             articleUpdated.setCodigo(a.getCodigo());
             articleUpdated.setDescripcion(a.getDescripcion());
             articleUpdated.setPrecio(a.getPrecio());
 
-            return Optional.of(articleRepository.save(articleUpdated));
+            return articleRepository.save(articleUpdated);
 
         } else {
             throw new ArticleNotFoundException("Article ID  " + id + "no encontrado.");
         }
     }
 
+    //todo que mensaje??
     @Override
-    public Article saveArticle(Article a) {
-        return articleRepository.save(a);
+    public Article saveArticle(Article article) {
+        try {
+            return articleRepository.save(article);
+        } catch (Exception e) {
+            throw new NotSavedException("Error saving article" + e.getMessage());
+        }
     }
 
     @Override
@@ -60,13 +78,12 @@ public class ArticleServiceImpl implements ArticleService {
 
     @Override
     public boolean deleteArticleById(long id) {
-        try {
-            Optional<Article> articleFound = articleRepository.findById(id);
+        Optional<Article> articleFound = articleRepository.findById(id);
+        if (articleFound.isPresent()) {
             articleRepository.delete(articleFound.get());
             return true;
-
-        } catch (Exception e) {
-            throw new ArticleNotFoundException("Articles not Found. ");
+        } else {
+            throw new ArticleNotFoundException("Article not found with id: " + id);
         }
     }
 }
